Crate lunar_lua

Safe bindings to lua.

Like lua, lunar_lua aimes to be extensible. For this reson all access to the lua api is provided through three traits:

LuaCore

LuaStandardLib

LuaAuxLib

each trait exposses functions from thier respective c header.

use lunar_lua::{State, LuaCore};

fn main() {
    let lua = State::new();
    lua.push(13f64);
    assert_eq!(13f64, lua.to_number(-1));
}

Custom State

Creating a custom state is as easy as implemanting the LuaConn trait.

use lunar_lua::{lua_aux::aux_new_state, ffi::lua_State,LuaConn, LuaCore, LuaConnection};
struct State(*mut lua_State);
unsafe impl LuaConn for State {
 fn get_conn(&self) -> LuaConnection<'_> {
    unsafe {
    self.0.as_ref().expect("valid ptr expected").into()
    }
 }
}
impl Drop for State {
    fn drop(&mut self) {
        unsafe {
            self.get_conn().borrow().close_conn();
        }
    }
}
fn main() {
    let lua = State(aux_new_state());
    lua.push(13f64);
    assert_eq!(13f64, lua.to_number(-1));
}

Re-exports

pub use lua_aux::LuaAuxLib;

pub use lua_core::LuaCore;

pub use lua_lib::LuaStandardLib;

pub use wrapper::*;

Modules

ffi

Raw bindings to lua

lua_aux

lua_core

lua_lib

wrapper

Default lua wrapper.

Macros

check_for_err

lua_error

Formats and reports an error. Calls aux_error.

Structs

LuaConnection

LuaStateRef

Traits

LuaConn