Module lunar_lua::lua_core

Enums

LuaStackValue

a container for valid stack types

Traits

LuaCore

Functions

new_state

Creates a new independent state and returns its main thread. Returns NULL if it cannot create the state (due to lack of memory). The argument f is the allocator function; Lua will do all memory allocation for this state through this function (see lua_Alloc). The second argument, ud, is an opaque pointer that Lua passes to the allocator in every call.