Trait lunar_lua::lua_core::LuaCore

pub trait LuaCore: LuaConn {

    // Provided methods
    fn pcall(
        &self,
        nargs: i32,
        nresults: i32,
        msgh: i32
    ) -> Result<(), LuaError> { ... }
    fn get_type(&self, index: i32) -> LuaType { ... }
    fn type_name(&self, index: i32) -> &str { ... }
    fn get_global(&self, name: &str) -> LuaType { ... }
    fn pop(&self, n: i32) { ... }
    fn set_top(&self, idx: i32) { ... }
    fn get_top(&self) -> i32 { ... }
    fn push_nil(&self) { ... }
    fn push_boolean(&self, b: bool) { ... }
    fn push_number(&self, n: f64) { ... }
    fn push_l_string<'a>(&'a self, s: &str, length: isize) -> &'a str { ... }
    fn push_string<'a>(&'a self, s: &str) -> &'a str { ... }
    fn push_value(&self, idx: i32) { ... }
    fn push<'a, 'lua, T: Into<LuaStackValue<'a>>>(
        &'lua self,
        value: T
    ) -> Option<&'lua str> { ... }
    fn is_nil(&self, idx: i32) -> bool { ... }
    fn is_boolean(&self, idx: i32) -> bool { ... }
    fn is_number(&self, idx: i32) -> bool { ... }
    fn is_string(&self, idx: i32) -> bool { ... }
    fn to_string(&self, idx: i32) -> &str { ... }
    fn to_number(&self, idx: i32) -> f64 { ... }
    fn to_boolean(&self, idx: i32) -> bool { ... }
    fn check_stack(&self, sz: i32) -> bool { ... }
    fn remove(&self, idx: i32) { ... }
    fn insert(&self, idx: i32) { ... }
    fn replace(&self, idx: i32) { ... }
    fn copy(&self, from_idx: i32, to_idx: i32) { ... }
    fn rotate(&self, idx: i32, n: i32) { ... }
    fn error(&self) { ... }

}

Provided Methods

fn pcall(&self, nargs: i32, nresults: i32, msgh: i32) -> Result<(), LuaError>

Calls a function (or a callable object) in protected mode.

Both nargs and nresults have the same meaning as in lua_call. If there are no errors during the call, lua_pcall behaves exactly like lua_call. However, if there is any error, pcall catches it, pushes a single value on the stack (the error object), and returns an error code. Like lua_call, lua_pcall always removes the function and its arguments from the stack.

If msgh is 0, then the error object returned on the stack is exactly the original error object. Otherwise, msgh is the stack index of a message handler. (This index cannot be a pseudo-index.) In case of runtime errors, this handler will be called with the error object and its return value will be the object returned on the stack by pcall.

Typically, the message handler is used to add more debug information to the error object, such as a stack traceback. Such information cannot be gathered after the return of pcall, since by then the stack has unwound.

fn get_type(&self, index: i32) -> LuaType

Returns the type of the value in the given valid index.

fn type_name(&self, index: i32) -> &str

Returns the name of the type encoded by the value tp, which must be one the values returned by get_type.

fn get_global(&self, name: &str) -> LuaType

Pushes onto the stack the value of the global name. Returns the type of that value.

fn pop(&self, n: i32)

Pops n elements from the stack.

fn set_top(&self, idx: i32)

Accepts any index, or 0, and sets the stack top to this index. If the new top is greater than the old one, then the new elements are filled with nil. If index is 0, then all stack elements are removed.

This function can run arbitrary code when removing an index marked as to-be-closed from the stack.

fn get_top(&self) -> i32

Returns the index of the top element in the stack. Because indices start at 1, this result is equal to the number of elements in the stack; in particular, 0 means an empty stack.

fn push_nil(&self)

Pushes a nil value onto the stack.

fn push_boolean(&self, b: bool)

Pushes a boolean value with value b onto the stack.

fn push_number(&self, n: f64)

Pushes a float with value n onto the stack.

fn push_l_string<'a>(&'a self, s: &str, length: isize) -> &'a str

Pushes the string pointed to by s with size len onto the stack. Lua will make or reuse an internal copy of the given string, so the memory at s can be freed or reused immediately after the function returns. The string can contain any binary data, including embedded zeros.

Returns a reference to the internal copy of the string (see 4.1.3).

fn push_string<'a>(&'a self, s: &str) -> &'a str

Pushes the zero-terminated string pointed to by s onto the stack. Lua will make or reuse an internal copy of the given string, so the memory at s can be freed or reused immediately after the function returns.

Returns a referetce to the internal copy of the string (see 4.1.3).

TODO If s is NULL, pushes nil and returns NULL.

fn push_value(&self, idx: i32)

Pushes a copy of the element at the given index onto the stack.

fn push<'a, 'lua, T: Into<LuaStackValue<'a>>>( &'lua self, value: T ) -> Option<&'lua str>

Pushes value onto stack

fn is_nil(&self, idx: i32) -> bool

Returns true if the value at the given index is nil, and false otherwise.

fn is_boolean(&self, idx: i32) -> bool

Returns true if the value at the given index is a boolean, and false otherwise.

fn is_number(&self, idx: i32) -> bool

Returns true if the value at the given index is a number or a string convertible to a number, and false otherwise.

fn is_string(&self, idx: i32) -> bool

Returns true if the value at the given index is a string or a number (which is always convertible to a string), and false otherwise.

fn to_string(&self, idx: i32) -> &str

Equivalent to lua_tolstring with len equal to NULL.

fn to_number(&self, idx: i32) -> f64

Equivalent to lua_tonumberx with isnum equal to NULL.

fn to_boolean(&self, idx: i32) -> bool

Converts the Lua value at the given index to a C boolean value (0 or 1). Like all tests in Lua, lua_toboolean returns true for any Lua value different from false and nil; otherwise it returns false. (If you want to accept only actual boolean values, use lua_isboolean to test the value’s type.)

fn check_stack(&self, sz: i32) -> bool

Ensures that the stack has space for at least n extra elements, that is, that you can safely push up to n values into it. It returns false if it cannot fulfill the request, either because it would cause the stack to be greater than a fixed maximum size (typically at least several thousand elements) or because it cannot allocate memory for the extra space. This function never shrinks the stack; if the stack already has space for the extra elements, it is left unchanged.

fn remove(&self, idx: i32)

Removes the element at the given valid index, shifting down the elements above this index to fill the gap. This function cannot be called with a pseudo-index, because a pseudo-index is not an actual stack position.

fn insert(&self, idx: i32)

Moves the top element into the given valid index, shifting up the elements above this index to open space. This function cannot be called with a pseudo-index, because a pseudo-index is not an actual stack position.

fn replace(&self, idx: i32)

Moves the top element into the given valid index without shifting any element (therefore replacing the value at that given index), and then pops the top element.

fn copy(&self, from_idx: i32, to_idx: i32)

Copies the element at index fromidx into the valid index toidx, replacing the value at that position. Values at other positions are not affected.

fn rotate(&self, idx: i32, n: i32)

Rotates the stack elements between the valid index idx and the top of the stack. The elements are rotated n positions in the direction of the top, for a positive n, or -n positions in the direction of the bottom, for a negative n. The absolute value of n must not be greater than the size of the slice being rotated. This function cannot be called with a pseudo-index, because a pseudo-index is not an actual stack position.

fn error(&self)

Raises a Lua error, using the value on the top of the stack as the error object. This function does a long jump, and therefore never returns (see aux_error).

Object Safety

This trait is not object safe.