Struct lunar_lua::state::State

pub struct State(/* private fields */);

Implementations

impl State

pub fn new() -> Self

Creates a new Lua state.

Panics:

out of memory

pub fn open_libs(&mut self)

Opens all standard Lua libraries into the given state.

pub fn load_buffer(&mut self, buff: &str, name: &str) -> Result<()>

Equivalent to load_buffer_with_mode with mode equal to None.

pub fn load_buffer_with_mode( &mut self, buff: &[u8], name: &str, mode: Option<LoadMode> ) -> Result<()>

Loads a buffer as a Lua chunk. This function uses lua_load to load the chunk in the buffer pointed to by buff

Params

name is the chunk name, used for debug information and error messages.

pub fn pcall(&mut self, nargs: u32, nresults: u32, msgh: u32) -> Result<()>

pub fn pop(&mut self, n: u32)

Pops n elements from the stack.

pub fn set_op(&mut self, idx: u32)

Accepts any index, or 0, and sets the stack top to this index. If the new top is greater than the old one, then the new elements are filled with nil. If index is 0, then all stack elements are removed.

This function can run arbitrary code when removing an index marked as to-be-closed from the stack.

pub fn to_string(&self, idx: i32) -> &str

pub fn get_op(&self) -> u32

Returns the index of the top element in the stack. Because indices start at 1, this result is equal to the number of elements in the stack; in particular, 0 means an empty stack.